# breadly-api/openapi.yaml
openapi: 3.1.0
info:
  title: Breadly API
  description: REST API contract for the Breadly application
  version: 1.0.0

servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /public/config:
    get:
      operationId: getPublicConfig
      summary: Get environment-specific configuration
      description: Returns public runtime configuration values (e.g. OIDC issuer and client ID). No authentication required.
      tags:
        - Public
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicConfig'
        '503':
          description: Configuration not available (environment variables not set)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /recipes:
    get:
      operationId: getRecipes
      summary: List all recipes
      tags:
        - Recipes
      responses:
        '200':
          description: A list of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'

    post:
      operationId: createRecipe
      summary: Create a new recipe
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRecipeDto'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /recipes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: MongoDB ObjectId of the recipe
        schema:
          type: string

    get:
      operationId: getRecipeById
      summary: Get a recipe by ID
      tags:
        - Recipes
      responses:
        '200':
          description: The recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      operationId: replaceRecipe
      summary: Fully replace a recipe
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRecipeDto'
      responses:
        '200':
          description: Recipe replaced successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: updateRecipe
      summary: Partially update a recipe
      tags:
        - Recipes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchRecipeDto'
      responses:
        '200':
          description: Recipe updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteRecipe
      summary: Delete a recipe
      tags:
        - Recipes
      responses:
        '204':
          description: Recipe deleted successfully
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /profile:
    get:
      operationId: getProfile
      summary: Get the authenticated user's profile
      description: Returns profile information derived from the authenticated user's JWT claims.
      tags:
        - Profile
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      operationId: getHealth
      summary: Get system health status
      description: Returns the health status of the API and its dependencies.
      tags:
        - Operations
      responses:
        '200':
          description: Health status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

components:
  schemas:
    Recipe:
      type: object
      required:
        - name
      properties:
        _id:
          type: string
          description: MongoDB ObjectId (string representation)
          readOnly: true
        name:
          type: string
          description: Name of the recipe
          minLength: 1

    CreateRecipeDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the recipe
          minLength: 1

    UpdateRecipeDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Name of the recipe
          minLength: 1

    PatchRecipeDto:
      type: object
      properties:
        name:
          type: string
          description: Name of the recipe
          minLength: 1

    Profile:
      type: object
      required:
        - sub
        - email
        - emailVerified
        - roles
      properties:
        sub:
          type: string
          description: Subject identifier (unique user ID)
        email:
          type: string
          description: User's email address
        emailVerified:
          type: boolean
          description: Whether the email address has been verified
        name:
          type: string
          description: User's full display name
        givenName:
          type: string
          description: User's first name
        familyName:
          type: string
          description: User's last name
        picture:
          type: string
          description: URL to the user's profile picture
        roles:
          type: array
          items:
            type: string
          description: List of roles assigned to the user

    HealthResponse:
      type: object
      required:
        - status
        - checks
      properties:
        status:
          type: string
          enum:
            - ok
            - degraded
          description: Overall system health status
        checks:
          type: object
          required:
            - api
            - database
          properties:
            api:
              $ref: '#/components/schemas/HealthCheck'
            database:
              $ref: '#/components/schemas/HealthCheck'

    HealthCheck:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
            - degraded
          description: Status of this individual check
        responseTime:
          type: string
          description: Response time of the check

    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human-readable error description
        statusCode:
          type: integer
          description: HTTP status code

    PublicConfig:
      type: object
      required:
        - idp
      properties:
        idp:
          type: object
          required:
            - issuer
            - clientId
          properties:
            issuer:
              type: string
              description: OIDC issuer URL
            clientId:
              type: string
              description: OIDC client ID